[feat] cli arg check

+++ convert.sh
@@ -1,5 +1,11 @@
 #!/bin/bash
-set -exo pipefail
+set -eo pipefail
+
+if [[ $# -ne 3 ]]; then
+  echo "Usage: $0 <input.pdf> <task_name> <output_dir>"
+  echo "Example: $0 presentation.pdf presentation output"
+  exit 1
+fi
 
 IMAGE_DIR=${2:-./Slides}
 rm -rf $IMAGE_DIR
